Websocket testing, polling test and close the server
I've created some websockets testing code
In the coming versions, this code will include a websocket server, such
that clients can communicate over an open websocket in javascript
instead of a tcp socket in php.

I've also bought a new device to get the current temperature. This
device requires polling, which was not enabled in my server. Still
learning on this topic and it does not work yet, but I'll get there.

Finally, you are now able to shut the openzwave-server down with the
EXIT command