All commands should respond with JSON now
All server responses are in JSON.
The messages sent to both websockets and TCP connections are nicely
formatted instead of the custom messages by the person who initially
created the TCP server.
Only messages to the server are still in this weird format. Not sure yet
if I will change this.
Let me know if anything is unstable!