Split the Polymer client over multiple files
The index file is readable again.
All style content is moved to the css file, all javascript content is
moved to the js file.
All html imports needed to run the app have also been moved to a
separate file.
Now the index consists purely of the html structure of the app.