Polymer Client mobile support
It was difficult to work with the Polymer Client on mobile devices,
because I didn't specify the viewport.
Now I have specified it and to make everything look good, I had to move
some things around.
The input for specific commands is moved to the menu.
The temperature gauges didn't fit, so they now wrap when they don't fit
on the screen.